code added

+++ 05. Input-Output in Java/02_ReadShortValue.java
@@ -5,7 +5,7 @@ public class ReadShortValue {
         Scanner sc = new Scanner(System.in);
         
         System.out.print("Enter a number: ");
-        short a = sc.nextByte();
+        short a = sc.nextShort();
         
         System.out.println("The number is: " + a);
     }
